pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "wine-review-chatloop"
version = "0.1.0"
description = ""
authors = [
    {name = "Saurabh Kamboj",email = "skamboj97@gmail.com"}
]
readme = "README.md"
requires-python = "<3.14,>=3.10"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "openai (>=1.108.0,<2.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "mem0ai (>=1.0.2,<2.0.0)",
]

[tool.poetry.scripts]
wine-chat = "cli.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
